* device/include/string.h: added size_t typedef, changed prototypes to use size_t, eliminated separate reentrant and non-reentrant declarations, added _memmove declaration * device/lib/_memcpy.c: changed to use size_t instead of int, changed /4 to >>2 to avoid division library call * device/lib/_memcmp.c, * device/lib/_memset.c, * device/lib/_strncat.c, * device/lib/_strncpy.c, * device/lib/_strncmp.c: changed to use size_t instead of int * device/lib/_memmove.c: new file (fixed bug #772294) * device/lib/Makefile.in: added _memmove.c * device/lib/z80/asm_strings.s: fixed bug #772290 * support/regression/tests/bitfields.c: attempt to fix host assertion failure on amd64-unknown-linux2.2
git-svn-id: https://sdcc.svn.sourceforge.net/svnroot/sdcc/trunk/sdcc@2881 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ device/lib/_memcmp.c
@@ -26,7 +26,7 @@
 int memcmp (
 	void * buf1,
 	void * buf2,
-	int count
+	size_t count
 	) 
 {
 	if (!count)
